+++ packages/frontend/src/types/DataSource.ts
@@ -1,3 +1,5 @@
+import { ErrorsAndControl } from './HookForm';
+
 export type FormCaption = {
   title?: string;
   description?: string;
@@ -5,4 +7,9 @@ export type FormCaption = {
 
 export type EditorWrapper = {
   gapBottom?: number;
+};
+
+export type Editors = ErrorsAndControl & {
+  title: string;
+  source: any;
 };
