+++ packages/frontend/src/types/React.ts
@@ -1,0 +1,3 @@
+export type ClassName = {
+  className: string;
+};
