+++ packages/frontend/src/types/index.ts
@@ -1,0 +1,2 @@
+export type { FormCaption, EditorWrapper, Editors } from './DataSource';
+export type { ErrorsAndControl } from './HookForm';
