+++ packages/frontend/src/view/DataSources/AddDataSource/AddDataSource.tsx
@@ -2,26 +2,25 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { Redirect } from 'react-router-dom';
 import { Controller } from 'react-hook-form';
-import clsx from 'clsx';
-import { JsonEditor as Editor } from 'jsoneditor-react';
-import 'jsoneditor-react/es/editor.min.css';
 
 import { ROUTES } from '../../../model/providers';
 import { useAddDataSource } from '../../../presenter/DataSources';
-import { WidgetRow, HugeWidget, WidgetHeader, WidgetBody, H6, Input, PrimaryButton } from '../../../ui';
-import { EditorWrapper } from './EditorWrapper';
-import { EditorCell } from './EditorCell';
+import {
+  WidgetRow,
+  HugeWidget,
+  WidgetHeader,
+  WidgetBody,
+  Input,
+  PrimaryButton,
+} from '../../../ui';
+import { EditorWrapper, EditorCell, Editors } from './Editors';
 import { FormCaption } from './FormCaption';
 import { AddDataSourceHeader } from './Header';
 import { useStyles } from './useStyles';
 
 export const AddDataSource: React.FC = () => {
   const { source, control, onSubmit, errors } = useAddDataSource();
-  const { visibleOverflow, editor, code, schema, editorHeader, editorErrorHeader } = useStyles({});
-
-  const editorClassName = clsx(editor, code);
-  const schemaClassName = clsx(editor, schema);
-  const editorConnectorHeader = clsx(editor, !!(errors && errors.handler) && editorErrorHeader);
+  const { visibleOverflow } = useStyles({});
 
   if (!source) return <Redirect to={ROUTES.DATA_SOURCES} />;
 
@@ -52,44 +51,13 @@ export const AddDataSource: React.FC = () => {
                 </EditorCell>
                 <EditorCell />
               </EditorWrapper>
-              <EditorWrapper>
-                <EditorCell>
-                  <H6 className={editorConnectorHeader}>Connector config:</H6>
-                </EditorCell>
-                <EditorCell>
-                  <H6 className={editorHeader}>Connector schema:</H6>
-                </EditorCell>
-              </EditorWrapper>
-              <EditorWrapper gapBottom={4}>
-                <EditorCell>
-                  <Controller
-                    control={control}
-                    name="handler"
-                    render={(props) => (
-                      <Editor
-                        htmlElementProps={{
-                          className: editorClassName,
-                        }}
-                        schema={source}
-                        navigationBar={false}
-                        mode="code"
-                        {...props}
-                      />
-                    )}
-                  />
-                </EditorCell>
-                <EditorCell>
-                  <Editor
-                    value={source}
-                    name={`${title} schema`}
-                    htmlElementProps={{
-                      className: schemaClassName,
-                    }}
-                    mode="view"
-                    navigationBar={false}
-                  />
-                </EditorCell>
-              </EditorWrapper>
+              <Editors
+                control={control}
+                source={source}
+                title={title}
+                errors={errors}
+              />
+
               <PrimaryButton type="submit">Save</PrimaryButton>
             </form>
           </WidgetBody>
