--- packages/frontend/src/view/DataSources/AddDataSource/EditorCell.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import { useStyles } from './useStyles';
-
-export const EditorCell:React.FC = ({ children }) => {
-  const { editorCell } = useStyles({});
-  return (
-    <div className={editorCell}>
-      {children}
-    </div>
-  );
-}
