+++ packages/frontend/src/view/DataSources/AddDataSource/Editors/EditorWrapper.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
+import { EditorWrapper as Props } from '../../../../types';
 import { useStyles } from './useStyles';
-import { EditorWrapper as Props } from './types';
 
-export const EditorWrapper:React.FC<Props> = ({ children, gapBottom }) => {
+export const EditorWrapper: React.FC<Props> = ({ children, gapBottom }) => {
   const { editorsWrapper } = useStyles({ gapBottom });
   return <section className={editorsWrapper}>{children}</section>;
-}
+};
