+++ packages/frontend/src/view/DataSources/AddDataSource/Editors/index.ts
@@ -1,0 +1,3 @@
+export { Editors } from './Editors';
+export { EditorCell } from './EditorCell';
+export { EditorWrapper } from './EditorWrapper';
